# lib/api-zod/src/generated/types/aiTestInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AiTestInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  message: string;
}