# lib/api-zod/src/generated/types/resumeGenerateInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResumeGenerateInput {
  /** @maxLength 200 */
  fullName: string;
  /** @maxLength 320 */
  email: string;
  /** @maxLength 80 */
  phone: string;
  /** @maxLength 200 */
  location: string;
  /** @maxLength 300 */
  school: string;
  /** @maxLength 300 */
  degree: string;
  /** @maxLength 100 */
  graduation: string;
  /** @maxLength 5000 */
  skills: string;
  /** @maxLength 10000 */
  projects: string;
  /** @maxLength 10000 */
  experience: string;
  /** @maxLength 5000 */
  certifications: string;
  /** @maxLength 5000 */
  achievements: string;
  /** @maxLength 200 */
  targetRole: string;
}